some more fixes

+++ src/MemoryHandler.h
@@ -23,7 +23,7 @@ namespace Memoryhandler {
 
     extern void setup_privileges (const TCHAR* psz_privilege, bool enable = true);
 
-    extern void create_memory (void **mem_ref, uint64_t size, uint32_t align);
+    extern void create_memory (void **mem_ref, uint64_t size, uint8_t align);
 
     extern void free_memory (void *mem_ref);
